@@ -23,33 +23,34 @@
     "update-coverage": "vitest tests/unit --run --coverage && npx istanbul-badges-readme"
   },
   "dependencies": {
-    "svelte": "^4.2.18"
+    "svelte": "^5.22.6"
   },
   "devDependencies": {
-    "@playwright/test": "1.44.1",
-    "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/kit": "^2.5.10",
-    "@sveltejs/package": "^2.3.1",
-    "@sveltejs/vite-plugin-svelte": "^3.1.1",
-    "@vitest/coverage-v8": "^1.6.0",
-    "eslint": "^9.4.0",
-    "eslint-plugin-svelte": "^2.39.3",
-    "hastscript": "^9.0.0",
-    "jsdom": "^24.1.0",
-    "mdsvex": "^0.11.2",
-    "mdsvexamples": "^0.4.1",
-    "prettier": "^3.3.2",
-    "prettier-plugin-svelte": "^3.2.4",
+    "@playwright/test": "1.51.0",
+    "@stylistic/eslint-plugin": "^4.2.0",
+    "@sveltejs/adapter-static": "^3.0.8",
+    "@sveltejs/kit": "^2.19.0",
+    "@sveltejs/package": "^2.3.10",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@vitest/coverage-v8": "^3.0.8",
+    "eslint": "^9.22.0",
+    "eslint-plugin-svelte": "^3.1.0",
+    "hastscript": "^9.0.1",
+    "jsdom": "^26.0.0",
+    "mdsvex": "^0.12.3",
+    "mdsvexamples": "^0.5.0",
+    "prettier": "^3.5.3",
+    "prettier-plugin-svelte": "^3.3.3",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-slug": "^6.0.0",
-    "svelte-check": "^3.8.0",
-    "svelte-preprocess": "^5.1.4",
-    "svelte-zoo": "^0.4.10",
-    "svelte2tsx": "^0.7.9",
-    "typescript": "5.4.5",
-    "typescript-eslint": "^7.13.0",
-    "vite": "^5.2.13",
-    "vitest": "^1.6.0"
+    "svelte-check": "^4.1.5",
+    "svelte-preprocess": "^6.0.3",
+    "svelte-zoo": "^0.4.17",
+    "svelte2tsx": "^0.7.35",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^8.26.0",
+    "vite": "^6.2.1",
+    "vitest": "^3.0.8"
   },
   "keywords": [
     "svelte",
