@@ -1,6 +1,12 @@
 <script lang="ts">
   import { name } from '$root/package.json'
   import { DemoNav } from '$site'
+  import type { Snippet } from 'svelte'
+
+  interface Props {
+    children?: Snippet;
+  }
+  let { children }: Props = $props();
 </script>
 
 <h1>
@@ -9,7 +15,7 @@
 
 <DemoNav style="place-content: center;" />
 
-<slot />
+{@render children?.()}
 
 <style>
   h1 {
